requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
psycopg2-binary==2.9.10
openai==1.78.0
langchain==0.3.25
langchain-ollama==0.3.2
pydantic==2.11.4
python-multipart==0.0.20
pgvector==0.3.6  # Required for PostgreSQL vector operations
